refactor(changelog-generator): extract `git()` into git.js
Again, just a little bit of decluttering.
